Move jest-matchers types file into `src`.
